+++ source/Net/Bazzline/Component/ProxyLogger/Event/BufferEvent.php
@@ -7,6 +7,7 @@
 namespace Net\Bazzline\Component\ProxyLogger\Event;
 
 use Net\Bazzline\Component\ProxyLogger\LogRequest\LogRequestBufferInterface;
+use Net\Bazzline\Component\ProxyLogger\LogRequest\LogRequestBufferAwareInterface;
 
 /**
  * Class BufferEvent
@@ -15,7 +16,7 @@ use Net\Bazzline\Component\ProxyLogger\LogRequest\LogRequestBufferInterface;
  * @author stev leibelt <artodeto@arcor.de>
  * @since 2013-11-08
  */
-class BufferEvent extends ProxyEvent
+class BufferEvent extends ProxyEvent implements LogRequestBufferAwareInterface
 {
     const ADD_LOG_REQUEST_TO_BUFFER = 'netBazzlineComponentProxyLoggerEvent.bufferEvent.addLogRequestToBuffer';
     const BUFFER_CLEAN = 'netBazzlineComponentProxyLoggerEvent.bufferEvent.bufferClean';
@@ -66,6 +67,16 @@ class BufferEvent extends ProxyEvent
     public function getLogRequestBuffer()
     {
         return $this->logRequestBuffer;
+    }
+
+    /**
+     * @return bool
+     * @author stev leibelt <artodeto@arcor.de>
+     * @since 2013-11-18
+     */
+    public function hasLogRequestBuffer()
+    {
+        return (!is_null($this->logRequestBuffer));
     }
 
     /**