+++ documentation/VersionHistory.md
@@ -5,6 +5,7 @@
     * added AbstractLogger
     * implement an event based handling for buffer logger to easy up creation of future buffer manipulators
     * moved files in component root to same hierarchy level like the loggers
+    * removed "LogRequestBufferFactory" in "BufferLogger" (buffer is now only needed in the event)
     * renamed namespace "proxy" to "logger"
 * [1.1.0](https://github.com/stevleibelt/php_component_proxy_logger/tree/1.1.0) - released at 2013-11-02
     * add example for using two manipulate buffer loggers in a proxy logger to act on different log levels (sending mail or sending sms/wake up call)